refactor($RacingGameController): refactor `RacingGameController`
- delete `GameConroller` and move `run()` to here
- use `RacingGame` in input car name logic and start game logic
